Fix hadoop-provided profile.
When this profile is enabled, it causes external/flume to fail during test
compilation. This is because an older version of io.netty gets included. The
problem is related to the fact that netty jars are not excluded from the hadoop
jars used in this profile. They are done correctly in the default mode.

To fix this, we avoid redefining the dependencies again for the profile. Instead
simply set a variable called hadoop.scope. This is also the approach used to
clean up pom.xml in 1.3.0 release.
https://github.com/apache/spark/pull/2982

We can get rid off this commit when we upgrade to 1.3.